@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.4.0'
     ext.protobufVersion = '0.8.12'
     ext.navVersion = "2.2.2"
 
@@ -47,7 +47,7 @@ subprojects {
     }
 
     ktlint {
-        debug = true
+        debug = false
         ignoreFailures = false
         coloredOutput = false
     }
