@@ -2,13 +2,21 @@ package de.rki.coronawarnapp.service.submission
 
 import de.rki.coronawarnapp.exception.NoGUIDOrTANSetException
 import de.rki.coronawarnapp.exception.NoRegistrationTokenSetException
-import de.rki.coronawarnapp.http.WebRequestBuilder
-import de.rki.coronawarnapp.http.playbook.BackgroundNoise
+import de.rki.coronawarnapp.playbook.BackgroundNoise
+import de.rki.coronawarnapp.playbook.Playbook
 import de.rki.coronawarnapp.storage.LocalData
+import de.rki.coronawarnapp.storage.SubmissionRepository
+import de.rki.coronawarnapp.submission.Symptoms
 import de.rki.coronawarnapp.transaction.SubmitDiagnosisKeysTransaction
+import de.rki.coronawarnapp.util.di.AppInjector
+import de.rki.coronawarnapp.util.di.ApplicationComponent
 import de.rki.coronawarnapp.util.formatter.TestResult
+import de.rki.coronawarnapp.verification.server.VerificationKeyType
+import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.matchers.shouldBe
 import io.mockk.MockKAnnotations
 import io.mockk.Runs
+import io.mockk.clearAllMocks
 import io.mockk.coEvery
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
@@ -16,26 +24,30 @@ import io.mockk.just
 import io.mockk.mockkObject
 import io.mockk.verify
 import kotlinx.coroutines.runBlocking
-import org.hamcrest.CoreMatchers.equalTo
-import org.hamcrest.MatcherAssert.assertThat
-import org.junit.Before
-import org.junit.Test
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
 
 class SubmissionServiceTest {
+
     private val guid = "123456-12345678-1234-4DA7-B166-B86D85475064"
     private val registrationToken = "asdjnskjfdniuewbheboqudnsojdff"
+    private val testResult = TestResult.PENDING
 
-    @MockK
-    private lateinit var webRequestBuilder: WebRequestBuilder
+    @MockK lateinit var backgroundNoise: BackgroundNoise
+    @MockK lateinit var mockPlaybook: Playbook
+    @MockK lateinit var appComponent: ApplicationComponent
 
-    @MockK
-    private lateinit var backgroundNoise: BackgroundNoise
+    private val symptoms = Symptoms(Symptoms.StartOf.OneToTwoWeeksAgo, Symptoms.Indication.POSITIVE)
 
-    @Before
+    @BeforeEach
     fun setUp() {
         MockKAnnotations.init(this)
-        mockkObject(WebRequestBuilder.Companion)
-        every { WebRequestBuilder.getInstance() } returns webRequestBuilder
+
+        mockkObject(AppInjector)
+        every { AppInjector.component } returns appComponent
+
+        every { appComponent.playbook } returns mockPlaybook
 
         mockkObject(BackgroundNoise.Companion)
         every { BackgroundNoise.getInstance() } returns backgroundNoise
@@ -43,14 +55,22 @@ class SubmissionServiceTest {
         mockkObject(SubmitDiagnosisKeysTransaction)
         mockkObject(LocalData)
 
+        mockkObject(SubmissionRepository)
+        every { SubmissionRepository.updateTestResult(any()) } just Runs
+
         every { LocalData.teletan() } returns null
         every { LocalData.testGUID() } returns null
         every { LocalData.registrationToken() } returns null
     }
 
-    @Test(expected = NoGUIDOrTANSetException::class)
-    fun registerDeviceWithoutTANOrGUIDFails() {
-        runBlocking {
+    @AfterEach
+    fun cleanUp() {
+        clearAllMocks()
+    }
+
+    @Test
+    fun registerDeviceWithoutTANOrGUIDFails(): Unit = runBlocking {
+        shouldThrow<NoGUIDOrTANSetException> {
             SubmissionService.asyncRegisterDevice()
         }
     }
@@ -64,8 +84,10 @@ class SubmissionServiceTest {
         every { LocalData.devicePairingSuccessfulTimestamp(any()) } just Runs
 
         coEvery {
-            webRequestBuilder.asyncGetRegistrationToken(any(), KeyType.GUID)
-        } returns registrationToken
+            mockPlaybook.initialRegistration(any(), VerificationKeyType.GUID)
+        } returns (registrationToken to TestResult.PENDING)
+        coEvery { mockPlaybook.testResult(registrationToken) } returns testResult
+
         every { backgroundNoise.scheduleDummyPattern() } just Runs
 
         runBlocking {
@@ -77,6 +99,7 @@ class SubmissionServiceTest {
             LocalData.devicePairingSuccessfulTimestamp(any())
             LocalData.testGUID(null)
             backgroundNoise.scheduleDummyPattern()
+            SubmissionRepository.updateTestResult(testResult)
         }
     }
 
@@ -89,8 +112,10 @@ class SubmissionServiceTest {
         every { LocalData.devicePairingSuccessfulTimestamp(any()) } just Runs
 
         coEvery {
-            webRequestBuilder.asyncGetRegistrationToken(any(), KeyType.TELETAN)
-        } returns registrationToken
+            mockPlaybook.initialRegistration(any(), VerificationKeyType.TELETAN)
+        } returns (registrationToken to TestResult.PENDING)
+        coEvery { mockPlaybook.testResult(registrationToken) } returns testResult
+
         every { backgroundNoise.scheduleDummyPattern() } just Runs
 
         runBlocking {
@@ -102,12 +127,13 @@ class SubmissionServiceTest {
             LocalData.devicePairingSuccessfulTimestamp(any())
             LocalData.teletan(null)
             backgroundNoise.scheduleDummyPattern()
+            SubmissionRepository.updateTestResult(testResult)
         }
     }
 
-    @Test(expected = NoRegistrationTokenSetException::class)
-    fun requestTestResultWithoutRegistrationTokenFails() {
-        runBlocking {
+    @Test
+    fun requestTestResultWithoutRegistrationTokenFails(): Unit = runBlocking {
+        shouldThrow<NoRegistrationTokenSetException> {
             SubmissionService.asyncRequestTestResult()
         }
     }
@@ -115,27 +141,33 @@ class SubmissionServiceTest {
     @Test
     fun requestTestResultSucceeds() {
         every { LocalData.registrationToken() } returns registrationToken
-        coEvery { webRequestBuilder.asyncGetTestResult(registrationToken) } returns TestResult.NEGATIVE.value
+        coEvery { mockPlaybook.testResult(registrationToken) } returns TestResult.NEGATIVE
 
         runBlocking {
-            assertThat(SubmissionService.asyncRequestTestResult(), equalTo(TestResult.NEGATIVE))
+            SubmissionService.asyncRequestTestResult() shouldBe TestResult.NEGATIVE
         }
     }
 
-    @Test(expected = NoRegistrationTokenSetException::class)
-    fun submitExposureKeysWithoutRegistrationTokenFails() {
-        runBlocking {
-            SubmissionService.asyncSubmitExposureKeys(listOf())
+    @Test
+    fun submitExposureKeysWithoutRegistrationTokenFails(): Unit = runBlocking {
+        shouldThrow<NoRegistrationTokenSetException> {
+            SubmissionService.asyncSubmitExposureKeys(listOf(), symptoms)
         }
     }
 
     @Test
     fun submitExposureKeysSucceeds() {
         every { LocalData.registrationToken() } returns registrationToken
-        coEvery { SubmitDiagnosisKeysTransaction.start(registrationToken, any()) } just Runs
+        coEvery {
+            SubmitDiagnosisKeysTransaction.start(
+                registrationToken,
+                any(),
+                symptoms
+            )
+        } just Runs
 
         runBlocking {
-            SubmissionService.asyncSubmitExposureKeys(listOf())
+            SubmissionService.asyncSubmitExposureKeys(listOf(), symptoms)
         }
     }
 
@@ -150,28 +182,5 @@ class SubmissionServiceTest {
             LocalData.registrationToken(null)
             LocalData.devicePairingSuccessfulTimestamp(0L)
         }
-    }
-
-    @Test
-    fun containsValidGUID() {
-        // valid
-        assertThat(
-            SubmissionService.containsValidGUID("https://bs-sd.de/covid-19/?$guid"),
-            equalTo(true)
-        )
-
-        // invalid
-        assertThat(
-            SubmissionService.containsValidGUID("https://no-guid-here"),
-            equalTo(false)
-        )
-    }
-
-    @Test
-    fun extractGUID() {
-        assertThat(
-            SubmissionService.extractGUID("https://bs-sd.de/covid-19/?$guid"),
-            equalTo(guid)
-        )
     }
 }
