@@ -1,13 +1,13 @@
 package de.rki.coronawarnapp.risk
 
 import com.google.android.gms.common.api.ApiException
-import de.rki.coronawarnapp.BuildConfig
 import de.rki.coronawarnapp.CoronaWarnApplication
 import de.rki.coronawarnapp.exception.ExceptionCategory
 import de.rki.coronawarnapp.exception.reporting.report
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationClient
 import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.storage.tracing.TracingIntervalRepository
+import de.rki.coronawarnapp.util.CWADebug
 import de.rki.coronawarnapp.util.TimeAndDateExtensions.daysToMilliseconds
 import de.rki.coronawarnapp.util.TimeAndDateExtensions.roundUpMsToDays
 
@@ -34,7 +34,7 @@ object TimeVariables {
      * The maximal runtime of a transaction
      * In milliseconds
      */
-    private const val TRANSACTION_TIMEOUT = 60000L
+    private const val TRANSACTION_TIMEOUT = 180000L
 
     /**
      * Getter function for [TRANSACTION_TIMEOUT]
@@ -99,20 +99,15 @@ object TimeVariables {
      * Delay in milliseconds for manual key retrieval process
      * Value for testing: 1 min =  1000 * 60 * 1
      * Value: 24 hours = 1000 * 60 * 60 * 24 milliseconds
+     *
+     * @return delay of key retrieval in milliseconds
      */
-    private val MANUAL_KEY_RETRIEVAL_DELAY =
-        if (BuildConfig.FLAVOR == "deviceForTesters") {
+    fun getManualKeyRetrievalDelay() =
+        if (CWADebug.buildFlavor == CWADebug.BuildFlavor.DEVICE_FOR_TESTERS) {
             MILISECONDS_IN_A_SECOND * SECONDS_IN_A_MINUTES
         } else {
             MILISECONDS_IN_A_SECOND * SECONDS_IN_A_MINUTES * MINUTES_IN_AN_HOUR * HOURS_IN_AN_DAY
         }
-
-    /**
-     * Getter function for [MANUAL_KEY_RETRIEVAL_DELAY]
-     *
-     * @return delay of key retrieval in milliseconds
-     */
-    fun getManualKeyRetrievalDelay() = MANUAL_KEY_RETRIEVAL_DELAY
 
     /**
      * This is the maximum attenuation duration value for the risk level calculation
@@ -175,19 +170,23 @@ object TimeVariables {
      * @return in milliseconds
      */
     fun getTimeActiveTracingDuration(): Long = System.currentTimeMillis() -
-            (getInitialExposureTracingActivationTimestamp() ?: 0L) -
-            LocalData.totalNonActiveTracing()
+        (getInitialExposureTracingActivationTimestamp() ?: 0L) -
+        LocalData.totalNonActiveTracing()
 
     suspend fun getActiveTracingDaysInRetentionPeriod(): Long {
         // the active tracing time during the retention period - all non active tracing times
         val tracingActiveMS = getTimeRangeFromRetentionPeriod()
+        val retentionPeriodInMS = getDefaultRetentionPeriodInMS()
+        val lastNonActiveTracingTimestamp = LocalData.lastNonActiveTracingTimestamp()
+        val current = System.currentTimeMillis()
+        val retentionTimestamp = current - retentionPeriodInMS
         val inactiveTracingIntervals = TracingIntervalRepository
             .getDateRepository(CoronaWarnApplication.getAppContext())
             .getIntervals()
             .toMutableList()
 
         // by default the tracing is assumed to be activated
-        // if the API is reachable we set the value accordingly 
+        // if the API is reachable we set the value accordingly
         val enIsDisabled = try {
             !InternalExposureNotificationClient.asyncIsEnabled()
         } catch (e: ApiException) {
@@ -195,21 +194,23 @@ object TimeVariables {
             false
         }
 
-        if (enIsDisabled) {
-            val current = System.currentTimeMillis()
-            val lastTimeTracingWasNotActivated = minOf(
-                LocalData.lastNonActiveTracingTimestamp() ?: current,
-                current - tracingActiveMS
-            )
-
+        // lastNonActiveTracingTimestamp could be null when en is disabled
+        // it only gets updated when you turn the en back on
+        // if en is disabled and lastNonActiveTracingTimestamp != null, only then we add a pair to
+        // the inactive intervals list to account for the time of inactivity between the last time
+        // en was not active and now.
+        if (enIsDisabled && lastNonActiveTracingTimestamp != null) {
+            val lastTimeTracingWasNotActivated =
+                LocalData.lastNonActiveTracingTimestamp() ?: current
             inactiveTracingIntervals.add(Pair(lastTimeTracingWasNotActivated, current))
         }
-
-        val finalTracingMS = tracingActiveMS - inactiveTracingIntervals
-            .map { it.second - it.first }
+        val inactiveTracingMS = inactiveTracingIntervals
+            .map { it.second - maxOf(it.first, retentionTimestamp) }
             .sum()
 
-        return finalTracingMS.roundUpMsToDays()
+        // because we delete periods that are past 14 days but tracingActiveMS counts from first
+        // ever activation, there are edge cases where tracingActiveMS gets to be > 14 days
+        return (minOf(tracingActiveMS, retentionPeriodInMS) - inactiveTracingMS).roundUpMsToDays()
     }
 
     /****************************************************
