@@ -3,15 +3,21 @@ package de.rki.coronawarnapp.service.submission
 import com.google.android.gms.nearby.exposurenotification.TemporaryExposureKey
 import de.rki.coronawarnapp.exception.NoGUIDOrTANSetException
 import de.rki.coronawarnapp.exception.NoRegistrationTokenSetException
-import de.rki.coronawarnapp.http.WebRequestBuilder
-import de.rki.coronawarnapp.http.playbook.BackgroundNoise
-import de.rki.coronawarnapp.http.playbook.PlaybookImpl
+import de.rki.coronawarnapp.playbook.BackgroundNoise
+import de.rki.coronawarnapp.playbook.Playbook
 import de.rki.coronawarnapp.storage.LocalData
+import de.rki.coronawarnapp.storage.SubmissionRepository
+import de.rki.coronawarnapp.submission.Symptoms
 import de.rki.coronawarnapp.transaction.SubmitDiagnosisKeysTransaction
+import de.rki.coronawarnapp.util.di.AppInjector
 import de.rki.coronawarnapp.util.formatter.TestResult
+import de.rki.coronawarnapp.verification.server.VerificationKeyType
 import de.rki.coronawarnapp.worker.BackgroundWorkScheduler
 
 object SubmissionService {
+
+    private val playbook: Playbook
+        get() = AppInjector.component.playbook
 
     suspend fun asyncRegisterDevice() {
         val testGUID = LocalData.testGUID()
@@ -27,53 +33,46 @@ object SubmissionService {
     }
 
     private suspend fun asyncRegisterDeviceViaGUID(guid: String) {
-        val registrationToken =
-            PlaybookImpl(WebRequestBuilder.getInstance()).initialRegistration(
+        val (registrationToken, testResult) =
+            playbook.initialRegistration(
                 guid,
-                KeyType.GUID
+                VerificationKeyType.GUID
             )
 
         LocalData.registrationToken(registrationToken)
         deleteTestGUID()
+        SubmissionRepository.updateTestResult(testResult)
     }
 
     private suspend fun asyncRegisterDeviceViaTAN(tan: String) {
-        val registrationToken =
-            PlaybookImpl(WebRequestBuilder.getInstance()).initialRegistration(
+        val (registrationToken, testResult) =
+            playbook.initialRegistration(
                 tan,
-                KeyType.TELETAN
+                VerificationKeyType.TELETAN
             )
 
         LocalData.registrationToken(registrationToken)
         deleteTeleTAN()
+        SubmissionRepository.updateTestResult(testResult)
     }
 
-    suspend fun asyncSubmitExposureKeys(keys: List<TemporaryExposureKey>) {
+    suspend fun asyncSubmitExposureKeys(keys: List<TemporaryExposureKey>, symptoms: Symptoms) {
         val registrationToken =
             LocalData.registrationToken() ?: throw NoRegistrationTokenSetException()
-        SubmitDiagnosisKeysTransaction.start(registrationToken, keys)
+        SubmitDiagnosisKeysTransaction.start(registrationToken, keys, symptoms)
     }
 
     suspend fun asyncRequestTestResult(): TestResult {
         val registrationToken =
             LocalData.registrationToken() ?: throw NoRegistrationTokenSetException()
 
-        return PlaybookImpl(WebRequestBuilder.getInstance()).testResult(registrationToken)
+        return playbook.testResult(registrationToken)
     }
 
     fun containsValidGUID(scanResult: String): Boolean {
-        if (scanResult.length > SubmissionConstants.MAX_QR_CODE_LENGTH ||
-            scanResult.count { it == SubmissionConstants.GUID_SEPARATOR } != 1
-        )
-            return false
-
-        val potentialGUID = extractGUID(scanResult)
-
-        return !(potentialGUID.isEmpty() || potentialGUID.length > SubmissionConstants.MAX_GUID_LENGTH)
+        val scanResult = QRScanResult(scanResult)
+        return scanResult.isValid
     }
-
-    fun extractGUID(scanResult: String): String =
-        scanResult.substringAfterLast(SubmissionConstants.GUID_SEPARATOR, "")
 
     fun storeTestGUID(guid: String) = LocalData.testGUID(guid)
 
