@@ -1,15 +1,19 @@
 package de.rki.coronawarnapp.transaction
 
 import com.google.android.gms.nearby.exposurenotification.TemporaryExposureKey
-import de.rki.coronawarnapp.http.WebRequestBuilder
-import de.rki.coronawarnapp.http.playbook.PlaybookImpl
+import de.rki.coronawarnapp.appconfig.AppConfigProvider
+import de.rki.coronawarnapp.appconfig.toNewConfig
+import de.rki.coronawarnapp.playbook.Playbook
+import de.rki.coronawarnapp.server.protocols.ApplicationConfigurationOuterClass.ApplicationConfiguration
 import de.rki.coronawarnapp.service.submission.SubmissionService
+import de.rki.coronawarnapp.submission.ExposureKeyHistoryCalculations
+import de.rki.coronawarnapp.submission.Symptoms
 import de.rki.coronawarnapp.transaction.SubmitDiagnosisKeysTransaction.SubmitDiagnosisKeysTransactionState.CLOSE
 import de.rki.coronawarnapp.transaction.SubmitDiagnosisKeysTransaction.SubmitDiagnosisKeysTransactionState.RETRIEVE_TAN_AND_SUBMIT_KEYS
 import de.rki.coronawarnapp.transaction.SubmitDiagnosisKeysTransaction.SubmitDiagnosisKeysTransactionState.RETRIEVE_TEMPORARY_EXPOSURE_KEY_HISTORY
 import de.rki.coronawarnapp.transaction.SubmitDiagnosisKeysTransaction.SubmitDiagnosisKeysTransactionState.STORE_SUCCESS
-import de.rki.coronawarnapp.util.ProtoFormatConverterExtensions.limitKeyCount
-import de.rki.coronawarnapp.util.ProtoFormatConverterExtensions.transformKeyHistoryToExternalFormat
+import de.rki.coronawarnapp.util.di.AppInjector
+import timber.log.Timber
 
 /**
  * The SubmitDiagnosisKeysTransaction is used to define an atomic Transaction for Key Reports. Its states allow an
@@ -36,6 +40,7 @@ import de.rki.coronawarnapp.util.ProtoFormatConverterExtensions.transformKeyHist
  */
 object SubmitDiagnosisKeysTransaction : Transaction() {
 
+    private const val FALLBACK_COUNTRY = "DE"
     override val TAG: String? = SubmitDiagnosisKeysTransaction::class.simpleName
 
     /** possible transaction states */
@@ -47,36 +52,60 @@ object SubmitDiagnosisKeysTransaction : Transaction() {
         CLOSE
     }
 
+    private val transactionScope: TransactionCoroutineScope by lazy {
+        AppInjector.component.transSubmitDiagnosisInjection.transactionScope
+    }
+
+    private val playbook: Playbook
+        get() = AppInjector.component.transSubmitDiagnosisInjection.playbook
+
+    private val appConfigProvider: AppConfigProvider
+        get() = AppInjector.component.transSubmitDiagnosisInjection.appConfigProvider
+
+    private val exposureKeyHistoryCalculations: ExposureKeyHistoryCalculations
+        get() = AppInjector.component.transSubmitDiagnosisInjection.exposureKeyHistoryCalculations
+
     /** initiates the transaction. This suspend function guarantees a successful transaction once completed. */
     suspend fun start(
         registrationToken: String,
-        keys: List<TemporaryExposureKey>
-    ) = lockAndExecuteUnique {
-        /****************************************************
-         * RETRIEVE TEMPORARY EXPOSURE KEY HISTORY
-         ****************************************************/
+        keys: List<TemporaryExposureKey>,
+        symptoms: Symptoms
+    ) = lockAndExecute(unique = true, scope = transactionScope) {
+
         val temporaryExposureKeyList = executeState(RETRIEVE_TEMPORARY_EXPOSURE_KEY_HISTORY) {
-                keys.limitKeyCount()
-                .transformKeyHistoryToExternalFormat()
+            exposureKeyHistoryCalculations.transformToKeyHistoryInExternalFormat(keys, symptoms)
         }
-        /****************************************************
-         * RETRIEVE TAN & SUBMIT KEYS
-         ****************************************************/
+
+        val visistedCountries =
+            appConfigProvider.getAppConfig().performSanityChecks().supportedCountriesList
+
         executeState(RETRIEVE_TAN_AND_SUBMIT_KEYS) {
-            PlaybookImpl(WebRequestBuilder.getInstance()).submission(
-                registrationToken,
-                temporaryExposureKeyList
+            val submissionData = Playbook.SubmissionData(
+                registrationToken = registrationToken,
+                temporaryExposureKeys = temporaryExposureKeyList,
+                consentToFederation = true,
+                visistedCountries = visistedCountries
             )
+            playbook.submission(submissionData)
         }
-        /****************************************************
-         * STORE SUCCESS
-         ****************************************************/
+
         executeState(STORE_SUCCESS) {
             SubmissionService.submissionSuccessful()
         }
-        /****************************************************
-         * CLOSE TRANSACTION
-         ****************************************************/
+
         executeState(CLOSE) {}
+    }
+
+    private fun ApplicationConfiguration.performSanityChecks(): ApplicationConfiguration {
+        var sanityChecked = this
+
+        if (sanityChecked.supportedCountriesList.isEmpty()) {
+            sanityChecked = sanityChecked.toNewConfig {
+                addSupportedCountries(FALLBACK_COUNTRY)
+            }
+            Timber.w("Country list was empty, corrected: %s", sanityChecked.supportedCountriesList)
+        }
+
+        return sanityChecked
     }
 }
