@@ -1,39 +1,22 @@
 package de.rki.coronawarnapp.ui.information
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
-import android.text.method.LinkMovementMethod
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentInformationAboutBinding
 import de.rki.coronawarnapp.ui.main.MainActivity
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * Basic Fragment which only displays static content.
  */
-class InformationAboutFragment : Fragment() {
-    companion object {
-        private val TAG: String? = InformationAboutFragment::class.simpleName
-    }
+class InformationAboutFragment : Fragment(R.layout.fragment_information_about) {
 
-    private var _binding: FragmentInformationAboutBinding? = null
-    private val binding: FragmentInformationAboutBinding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentInformationAboutBinding.inflate(inflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+    private val binding: FragmentInformationAboutBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -43,7 +26,11 @@ class InformationAboutFragment : Fragment() {
 
     private fun setLinks() {
         binding.informationAboutEasyLanguage
-            .movementMethod = LinkMovementMethod.getInstance()
+            .setOnClickListener {
+            val browserIntent = Intent(Intent.ACTION_VIEW,
+                Uri.parse(getString(R.string.onboarding_tracing_easy_language_explanation_url)))
+            startActivity(browserIntent)
+        }
     }
 
     override fun onResume() {
