@@ -2,38 +2,21 @@ package de.rki.coronawarnapp.ui.information
 
 import android.os.Bundle
 import android.text.method.LinkMovementMethod
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentInformationLegalBinding
 import de.rki.coronawarnapp.ui.main.MainActivity
+import de.rki.coronawarnapp.util.convertToHyperlink
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * Basic Fragment which only displays static content.
  */
-class InformationLegalFragment : Fragment() {
-    companion object {
-        private val TAG: String? = InformationLegalFragment::class.simpleName
-    }
+class InformationLegalFragment : Fragment(R.layout.fragment_information_legal) {
 
-    private var _binding: FragmentInformationLegalBinding? = null
-    private val binding: FragmentInformationLegalBinding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentInformationLegalBinding.inflate(inflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+    private val binding: FragmentInformationLegalBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -42,11 +25,11 @@ class InformationLegalFragment : Fragment() {
     }
 
     /**
-     * Make the links clickable
+     * Make the links clickable and convert to hyperlink
      */
     private fun setUpContactFormLinks() {
         binding.informationLegalContactForm.informationLegalContactForm
-            .movementMethod = LinkMovementMethod.getInstance()
+            .convertToHyperlink(getString(R.string.information_legal_subtitle_contact_url))
         binding.informationLegalContactForm.informationLegalContactFormNonEnDe
             .movementMethod = LinkMovementMethod.getInstance()
     }
