@@ -1,38 +1,19 @@
 package de.rki.coronawarnapp.ui.information
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentInformationPrivacyBinding
 import de.rki.coronawarnapp.ui.main.MainActivity
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * Basic Fragment which only displays static content.
  */
-class InformationPrivacyFragment : Fragment() {
-    companion object {
-        private val TAG: String? = InformationPrivacyFragment::class.simpleName
-    }
-
-    private var _binding: FragmentInformationPrivacyBinding? = null
-    private val binding: FragmentInformationPrivacyBinding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentInformationPrivacyBinding.inflate(inflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+class InformationPrivacyFragment : Fragment(R.layout.fragment_information_privacy) {
+    private val binding: FragmentInformationPrivacyBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
