@@ -10,16 +10,24 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.ViewModelProviders
 import androidx.lifecycle.lifecycleScope
+import dagger.android.AndroidInjector
+import dagger.android.DispatchingAndroidInjector
+import dagger.android.HasAndroidInjector
 import de.rki.coronawarnapp.R
+import de.rki.coronawarnapp.nearby.InternalExposureNotificationClient
+import de.rki.coronawarnapp.playbook.BackgroundNoise
 import de.rki.coronawarnapp.storage.LocalData
-import de.rki.coronawarnapp.http.playbook.BackgroundNoise
+import de.rki.coronawarnapp.ui.base.startActivitySafely
 import de.rki.coronawarnapp.ui.viewmodel.SettingsViewModel
+import de.rki.coronawarnapp.util.BackgroundPrioritization
 import de.rki.coronawarnapp.util.ConnectivityHelper
 import de.rki.coronawarnapp.util.DialogHelper
-import de.rki.coronawarnapp.util.ExternalActionHelper
-import de.rki.coronawarnapp.util.PowerManagementHelper
+import de.rki.coronawarnapp.util.device.PowerManagement
+import de.rki.coronawarnapp.util.di.AppInjector
 import de.rki.coronawarnapp.worker.BackgroundWorkScheduler
 import kotlinx.coroutines.launch
+import timber.log.Timber
+import javax.inject.Inject
 
 /**
  * This activity holds all the fragments (except onboarding) and also registers a listener for
@@ -29,7 +37,7 @@ import kotlinx.coroutines.launch
  * @see ConnectivityHelper
  * @see BackgroundWorkScheduler
  */
-class MainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity(), HasAndroidInjector {
     companion object {
         private val TAG: String? = MainActivity::class.simpleName
 
@@ -38,10 +46,19 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
+    @Inject lateinit var dispatchingAndroidInjector: DispatchingAndroidInjector<Any>
+    override fun androidInjector(): AndroidInjector<Any> = dispatchingAndroidInjector
+
     private val FragmentManager.currentNavigationFragment: Fragment?
         get() = primaryNavigationFragment?.childFragmentManager?.fragments?.first()
 
     private lateinit var settingsViewModel: SettingsViewModel
+
+    @Inject
+    lateinit var backgroundPrioritization: BackgroundPrioritization
+
+    @Inject
+    lateinit var powerManagement: PowerManagement
 
     /**
      * Register connection callback.
@@ -78,11 +95,15 @@ class MainActivity : AppCompatActivity() {
         }
 
         override fun onLocationUnavailable() {
-            settingsViewModel.updateLocationEnabled(false)
+            val canIgnoreLocationEnabled =
+                InternalExposureNotificationClient.deviceSupportsLocationlessScanning()
+            settingsViewModel.updateLocationEnabled(canIgnoreLocationEnabled)
+            Timber.d("Location unavailable but can be ignored? $canIgnoreLocationEnabled")
         }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        AppInjector.setup(this)
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         settingsViewModel = ViewModelProviders.of(this).get(SettingsViewModel::class.java)
@@ -117,7 +138,7 @@ class MainActivity : AppCompatActivity() {
             R.string.onboarding_energy_optimized_dialog_button_negative,
             false, {
                 // go to battery optimization
-                ExternalActionHelper.disableBatteryOptimizations(this)
+                startActivitySafely(powerManagement.disableBatteryOptimizationsIntent)
             }, {
                 // keep battery optimization enabled
                 showManualCheckingRequiredDialog()
@@ -126,7 +147,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun checkForEnergyOptimizedEnabled() {
-        if (!PowerManagementHelper.isIgnoringBatteryOptimizations(this)) {
+        if (!backgroundPrioritization.isBackgroundActivityPrioritized) {
             showEnergyOptimizedEnabledForBackground()
         }
     }
