@@ -1,9 +1,7 @@
 package de.rki.coronawarnapp.ui.main
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -11,40 +9,21 @@ import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentMainShareBinding
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
 import de.rki.coronawarnapp.util.ExternalActionHelper
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This fragment informs the user about what he is going to share and how he is going to help everybody with this :)
  *
  * @see TracingViewModel
  */
-class MainShareFragment : Fragment() {
-
-    companion object {
-        private val TAG: String? = MainShareFragment::class.simpleName
-    }
+class MainShareFragment : Fragment(R.layout.fragment_main_share) {
 
     private val tracingViewModel: TracingViewModel by activityViewModels()
-    private var _binding: FragmentMainShareBinding? = null
-    private val binding: FragmentMainShareBinding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentMainShareBinding.inflate(inflater)
-        binding.tracingViewModel = tracingViewModel
-        binding.lifecycleOwner = this
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+    private val binding: FragmentMainShareBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding.tracingViewModel = tracingViewModel
         setButtonOnClickListener()
     }
 
