@@ -1,14 +1,14 @@
 package de.rki.coronawarnapp.ui.onboarding
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.appcompat.app.AlertDialog
 import androidx.core.app.NotificationManagerCompat
 import androidx.fragment.app.Fragment
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentOnboardingNotificationsBinding
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This fragment ask the user if he wants to get notifications and finishes the onboarding afterwards.
@@ -16,27 +16,9 @@ import de.rki.coronawarnapp.databinding.FragmentOnboardingNotificationsBinding
  * @see NotificationManagerCompat
  * @see AlertDialog
  */
-class OnboardingNotificationsFragment : Fragment() {
-    companion object {
-        private val TAG: String? = OnboardingNotificationsFragment::class.simpleName
-    }
+class OnboardingNotificationsFragment : Fragment(R.layout.fragment_onboarding_notifications) {
 
-    private var _binding: FragmentOnboardingNotificationsBinding? = null
-    private val binding: FragmentOnboardingNotificationsBinding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentOnboardingNotificationsBinding.inflate(inflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+    private val binding: FragmentOnboardingNotificationsBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
