@@ -2,24 +2,22 @@ package de.rki.coronawarnapp.ui.onboarding
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentOnboardingTracingBinding
-import de.rki.coronawarnapp.exception.ExceptionCategory
-import de.rki.coronawarnapp.exception.reporting.report
-import de.rki.coronawarnapp.nearby.InternalExposureNotificationClient
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationPermissionHelper
-import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.ui.doNavigate
 import de.rki.coronawarnapp.util.DialogHelper
-import kotlinx.coroutines.launch
+import de.rki.coronawarnapp.util.di.AutoInject
+import de.rki.coronawarnapp.util.ui.observe2
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
+import de.rki.coronawarnapp.util.viewmodel.CWAViewModelFactoryProvider
+import de.rki.coronawarnapp.util.viewmodel.cwaViewModels
+import javax.inject.Inject
 
 /**
  * This fragment ask the user if he wants to enable tracing.
@@ -27,16 +25,14 @@ import kotlinx.coroutines.launch
  * @see InternalExposureNotificationPermissionHelper
  * @see AlertDialog
  */
-class OnboardingTracingFragment : Fragment(),
-    InternalExposureNotificationPermissionHelper.Callback {
-
-    companion object {
-        private val TAG: String? = OnboardingTracingFragment::class.simpleName
-    }
+class OnboardingTracingFragment : Fragment(R.layout.fragment_onboarding_tracing),
+    InternalExposureNotificationPermissionHelper.Callback, AutoInject {
 
     private lateinit var internalExposureNotificationPermissionHelper: InternalExposureNotificationPermissionHelper
-    private var _binding: FragmentOnboardingTracingBinding? = null
-    private val binding: FragmentOnboardingTracingBinding get() = _binding!!
+    private val binding: FragmentOnboardingTracingBinding by viewBindingLazy()
+
+    @Inject lateinit var viewModelFactory: CWAViewModelFactoryProvider.Factory
+    private val vm: OnboardingTracingFragmentViewModel by cwaViewModels { viewModelFactory }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         internalExposureNotificationPermissionHelper.onResolutionComplete(
@@ -51,29 +47,19 @@ class OnboardingTracingFragment : Fragment(),
             InternalExposureNotificationPermissionHelper(this, this)
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentOnboardingTracingBinding.inflate(inflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         setButtonOnClickListener()
+        vm.countryList.observe2(this) {
+            binding.countryData = it
+        }
+        vm.saveInteroperabilityUsed()
     }
 
     override fun onResume() {
         super.onResume()
         binding.onboardingTracingContainer.sendAccessibilityEvent(AccessibilityEvent.TYPE_ANNOUNCEMENT)
-        resetTracing()
+        vm.resetTracing()
     }
 
     private fun setButtonOnClickListener() {
@@ -84,7 +70,7 @@ class OnboardingTracingFragment : Fragment(),
             showCancelDialog()
         }
         binding.onboardingButtonBack.buttonIcon.setOnClickListener {
-            (activity as OnboardingActivity).goBack()
+            (requireActivity() as OnboardingActivity).goBack()
         }
     }
 
@@ -114,24 +100,5 @@ class OnboardingTracingFragment : Fragment(),
         findNavController().doNavigate(
             OnboardingTracingFragmentDirections.actionOnboardingTracingFragmentToOnboardingTestFragment()
         )
-    }
-
-    private fun resetTracing() {
-        // Reset tracing state in onboarding
-        lifecycleScope.launch {
-            try {
-                if (InternalExposureNotificationClient.asyncIsEnabled()) {
-                    InternalExposureNotificationClient.asyncStop()
-                    // Reset initial activation timestamp
-                    LocalData.initialTracingActivationTimestamp(0L)
-                }
-            } catch (exception: Exception) {
-                exception.report(
-                    ExceptionCategory.EXPOSURENOTIFICATION,
-                    OnboardingTracingFragment.TAG,
-                    null
-                )
-            }
-        }
     }
 }
