@@ -1,17 +1,17 @@
 package de.rki.coronawarnapp.ui.settings
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentSettingsBackgroundPriorityBinding
+import de.rki.coronawarnapp.ui.base.startActivitySafely
 import de.rki.coronawarnapp.ui.main.MainActivity
 import de.rki.coronawarnapp.ui.viewmodel.SettingsViewModel
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
-import de.rki.coronawarnapp.util.ExternalActionHelper
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This is the setting background priority page. Here the user sees the background priority setting status.
@@ -20,33 +20,15 @@ import de.rki.coronawarnapp.util.ExternalActionHelper
  * @see TracingViewModel
  * @see SettingsViewModel
  */
-class SettingsBackgroundPriorityFragment : Fragment() {
-    companion object {
-        private val TAG: String? = SettingsBackgroundPriorityFragment::class.simpleName
-    }
+class SettingsBackgroundPriorityFragment :
+    Fragment(R.layout.fragment_settings_background_priority) {
 
     private val settingsViewModel: SettingsViewModel by activityViewModels()
-    private var _binding: FragmentSettingsBackgroundPriorityBinding? = null
-    private val binding: FragmentSettingsBackgroundPriorityBinding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentSettingsBackgroundPriorityBinding.inflate(inflater)
-        binding.settingsViewModel = settingsViewModel
-        binding.lifecycleOwner = this
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+    private val binding: FragmentSettingsBackgroundPriorityBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding.settingsViewModel = settingsViewModel
         setButtonOnClickListener()
     }
 
@@ -54,7 +36,7 @@ class SettingsBackgroundPriorityFragment : Fragment() {
         super.onResume()
         binding.settingsBackgroundPriorityContainer.sendAccessibilityEvent(AccessibilityEvent.TYPE_ANNOUNCEMENT)
         // refresh required data
-        settingsViewModel.refreshBackgroundPriorityEnabled(requireContext())
+        settingsViewModel.refreshBackgroundPriorityEnabled()
     }
 
     private fun setButtonOnClickListener() {
@@ -66,14 +48,19 @@ class SettingsBackgroundPriorityFragment : Fragment() {
             it.setOnClickListener {
                 val isPriorityEnabled = settingsViewModel.isBackgroundPriorityEnabled.value == true
 
-                if (!isPriorityEnabled)
-                    ExternalActionHelper.disableBatteryOptimizations(requireContext())
+                if (!isPriorityEnabled) {
+                    (requireActivity() as MainActivity).apply {
+                        startActivitySafely(powerManagement.disableBatteryOptimizationsIntent)
+                    }
+                }
             }
         }
 
         // explanatory card
         binding.settingsTracingStatusConnection.tracingStatusCardButton.setOnClickListener {
-            ExternalActionHelper.toBatteryOptimizationSettings(requireContext())
+            (requireActivity() as MainActivity).apply {
+                startActivity(powerManagement.toBatteryOptimizationSettingsIntent)
+            }
         }
 
         // back navigation
