@@ -2,13 +2,12 @@ package de.rki.coronawarnapp.ui.settings
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentSettingsTracingBinding
 import de.rki.coronawarnapp.exception.ExceptionCategory
@@ -16,14 +15,15 @@ import de.rki.coronawarnapp.exception.reporting.report
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationClient
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationPermissionHelper
 import de.rki.coronawarnapp.storage.LocalData
+import de.rki.coronawarnapp.ui.doNavigate
 import de.rki.coronawarnapp.ui.main.MainActivity
 import de.rki.coronawarnapp.ui.viewmodel.SettingsViewModel
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
 import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.ExternalActionHelper
-import de.rki.coronawarnapp.util.IGNORE_CHANGE_TAG
-import de.rki.coronawarnapp.util.PowerManagementHelper
 import de.rki.coronawarnapp.util.formatter.formatTracingSwitchEnabled
+import de.rki.coronawarnapp.util.ui.observe2
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 import de.rki.coronawarnapp.worker.BackgroundWorkScheduler
 import kotlinx.coroutines.launch
 
@@ -35,7 +35,7 @@ import kotlinx.coroutines.launch
  * @see InternalExposureNotificationClient
  * @see InternalExposureNotificationPermissionHelper
  */
-class SettingsTracingFragment : Fragment(),
+class SettingsTracingFragment : Fragment(R.layout.fragment_settings_tracing),
     InternalExposureNotificationPermissionHelper.Callback {
 
     companion object {
@@ -44,30 +44,21 @@ class SettingsTracingFragment : Fragment(),
 
     private val tracingViewModel: TracingViewModel by activityViewModels()
     private val settingsViewModel: SettingsViewModel by activityViewModels()
-    private var _binding: FragmentSettingsTracingBinding? = null
-    private val binding: FragmentSettingsTracingBinding get() = _binding!!
+    private val binding: FragmentSettingsTracingBinding by viewBindingLazy()
 
     private lateinit var internalExposureNotificationPermissionHelper: InternalExposureNotificationPermissionHelper
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentSettingsTracingBinding.inflate(inflater)
-        binding.tracingViewModel = tracingViewModel
-        binding.settingsViewModel = settingsViewModel
-        binding.lifecycleOwner = this
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding.tracingViewModel = tracingViewModel
+        binding.settingsViewModel = settingsViewModel
+
+        tracingViewModel.navigateToInteroperability.observe2(this) {
+            if (it) {
+                navigateToInteroperability()
+            }
+        }
+
         setButtonOnClickListener()
     }
 
@@ -99,13 +90,15 @@ class SettingsTracingFragment : Fragment(),
         val switch = binding.settingsTracingSwitchRow.settingsSwitchRowSwitch
         val back = binding.settingsTracingHeader.headerButtonBack.buttonIcon
         val bluetooth = binding.settingsTracingStatusBluetooth.tracingStatusCardButton
-        val connection = binding.settingsTracingStatusConnection.tracingStatusCardButton
         val location = binding.settingsTracingStatusLocation.tracingStatusCardButton
+        val interoperability = binding.settingsInteroperabilityRow.settingsPlainRow
+
         internalExposureNotificationPermissionHelper =
             InternalExposureNotificationPermissionHelper(this, this)
-        switch.setOnCheckedChangeListener { _, _ ->
+        switch.setOnCheckedChangeListener { view, _ ->
+
             // Make sure that listener is called by user interaction
-            if (switch.tag != IGNORE_CHANGE_TAG) {
+            if (view.isPressed) {
                 startStopTracing()
                 // Focus on the body text after to announce the tracing status for accessibility reasons
                 binding.settingsTracingSwitchRow.settingsSwitchRowHeaderBody
@@ -117,15 +110,12 @@ class SettingsTracingFragment : Fragment(),
                 tracingViewModel.isTracingEnabled.value ?: throw IllegalArgumentException()
             val isBluetoothEnabled =
                 settingsViewModel.isBluetoothEnabled.value ?: throw IllegalArgumentException()
-            val isConnectionEnabled =
-                settingsViewModel.isConnectionEnabled.value ?: throw IllegalArgumentException()
             val isLocationEnabled =
                 settingsViewModel.isLocationEnabled.value ?: throw IllegalArgumentException()
             // check if the row is clickable, this adds the switch behaviour
             val isEnabled = formatTracingSwitchEnabled(
                 isTracingEnabled,
                 isBluetoothEnabled,
-                isConnectionEnabled,
                 isLocationEnabled
             )
             if (isEnabled) startStopTracing()
@@ -139,9 +129,17 @@ class SettingsTracingFragment : Fragment(),
         location.setOnClickListener {
             ExternalActionHelper.toMainSettings(requireContext())
         }
-        connection.setOnClickListener {
-            ExternalActionHelper.toConnections(requireContext())
+        interoperability.setOnClickListener {
+            tracingViewModel.onInteroperabilitySettingPressed()
         }
+    }
+
+    private fun navigateToInteroperability() {
+        findNavController()
+            .doNavigate(
+                SettingsTracingFragmentDirections
+                    .actionSettingsTracingFragmentToInteropCountryConfigurationFragment()
+            )
     }
 
     private fun startStopTracing() {
@@ -162,7 +160,8 @@ class SettingsTracingFragment : Fragment(),
                         // activated during onboarding
                         showConsentDialog()
                         // check if background processing is switched off, if it is, show the manual calculation dialog explanation before turning on.
-                        if (!PowerManagementHelper.isIgnoringBatteryOptimizations(requireActivity())) {
+                        val activity = requireActivity() as MainActivity
+                        if (!activity.backgroundPrioritization.isBackgroundActivityPrioritized) {
                             showManualCheckingRequiredDialog()
                         }
                     }
