@@ -1,9 +1,9 @@
 package de.rki.coronawarnapp.util
 
-import KeyExportFormat.SignatureInfo
+import de.rki.coronawarnapp.server.protocols.KeyExportFormat
 
 object SignatureHelper {
-    val clientSig: SignatureInfo = SignatureInfo.newBuilder()
+    val clientSig: KeyExportFormat.SignatureInfo = KeyExportFormat.SignatureInfo.newBuilder()
         .setAndroidPackage("de.rki.coronawarnapp")
         .setAppBundleId("de.rki.coronawarnapp")
         .setSignatureAlgorithm("ECDSA")
