@@ -8,8 +8,7 @@ object TracingStatusHelper {
     const val TRACING_ACTIVE = 0
     const val TRACING_INACTIVE = 1
     const val BLUETOOTH = 2
-    const val CONNECTION = 3
-    const val LOCATION = 4
+    const val LOCATION = 3
 }
 
 /**
@@ -18,19 +17,16 @@ object TracingStatusHelper {
  *
  * @param tracing
  * @param bluetooth
- * @param connection
  * @param location
  * @return Int
  */
-fun tracingStatusHelper(tracing: Boolean, bluetooth: Boolean, connection: Boolean, location: Boolean): Int {
+fun tracingStatusHelper(tracing: Boolean, bluetooth: Boolean, location: Boolean): Int {
     return if (!tracing) {
         TracingStatusHelper.TRACING_INACTIVE
     } else if (!location) {
         TracingStatusHelper.LOCATION
     } else if (!bluetooth) {
         TracingStatusHelper.BLUETOOTH
-    } else if (!connection) {
-        TracingStatusHelper.CONNECTION
     } else {
         TracingStatusHelper.TRACING_ACTIVE
     }
