@@ -20,45 +20,46 @@
 package de.rki.coronawarnapp.util.security
 
 import android.annotation.SuppressLint
-import android.content.Context
 import android.content.SharedPreferences
 import android.os.Build
 import android.util.Base64
-import androidx.security.crypto.EncryptedSharedPreferences
-import androidx.security.crypto.MasterKeys
-import de.rki.coronawarnapp.CoronaWarnApplication
+import androidx.annotation.VisibleForTesting
 import de.rki.coronawarnapp.exception.CwaSecurityException
+import de.rki.coronawarnapp.util.di.AppInjector
+import de.rki.coronawarnapp.util.di.ApplicationComponent
 import de.rki.coronawarnapp.util.security.SecurityConstants.CWA_APP_SQLITE_DB_PW
 import de.rki.coronawarnapp.util.security.SecurityConstants.DB_PASSWORD_MAX_LENGTH
 import de.rki.coronawarnapp.util.security.SecurityConstants.DB_PASSWORD_MIN_LENGTH
 import de.rki.coronawarnapp.util.security.SecurityConstants.ENCRYPTED_SHARED_PREFERENCES_FILE
+import timber.log.Timber
 import java.security.SecureRandom
 
 /**
  * Key Store and Password Access
  */
 object SecurityHelper {
-    private val keyGenParameterSpec = MasterKeys.AES256_GCM_SPEC
-    private val masterKeyAlias = MasterKeys.getOrCreate(keyGenParameterSpec)
 
-    val globalEncryptedSharedPreferencesInstance: SharedPreferences by lazy {
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    internal val encryptedPreferencesProvider: (ApplicationComponent) -> SharedPreferences = {
+        val factory = it.encryptedPreferencesFactory
+        val encryptionErrorResetTool = it.errorResetTool
         withSecurityCatch {
-            CoronaWarnApplication.getAppContext()
-                .getEncryptedSharedPrefs(ENCRYPTED_SHARED_PREFERENCES_FILE)
+            try {
+                factory.create(ENCRYPTED_SHARED_PREFERENCES_FILE)
+            } catch (e: Exception) {
+                if (encryptionErrorResetTool.tryResetIfNecessary(e)) {
+                    Timber.w("We could recovery from this error via reset. Now retrying.")
+                    factory.create(ENCRYPTED_SHARED_PREFERENCES_FILE)
+                } else {
+                    throw e
+                }
+            }
         }
     }
 
-    /**
-     * Initializes the private encrypted key store
-     */
-    private fun Context.getEncryptedSharedPrefs(fileName: String) = EncryptedSharedPreferences
-        .create(
-            fileName,
-            masterKeyAlias,
-            this,
-            EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
-            EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
-        )
+    val globalEncryptedSharedPreferencesInstance: SharedPreferences by lazy {
+        encryptedPreferencesProvider(AppInjector.component)
+    }
 
     private val String.toPreservedByteArray: ByteArray
         get() = Base64.decode(this, Base64.NO_WRAP)
