@@ -1,17 +1,23 @@
 package de.rki.coronawarnapp.util.security
 
-import KeyExportFormat
 import android.security.keystore.KeyProperties
 import android.util.Base64
-import de.rki.coronawarnapp.BuildConfig
+import de.rki.coronawarnapp.environment.EnvironmentSetup
+import de.rki.coronawarnapp.server.protocols.KeyExportFormat
 import timber.log.Timber
 import java.security.KeyFactory
 import java.security.Signature
 import java.security.spec.X509EncodedKeySpec
+import javax.inject.Inject
+import javax.inject.Singleton
 
-class VerificationKeys {
+@Singleton
+class VerificationKeys @Inject constructor(
+    private val environmentSetup: EnvironmentSetup
+) {
     companion object {
         private const val KEY_DELIMITER = ","
+        private val TAG = VerificationKeys::class.java.simpleName
     }
 
     private val keyFactory = KeyFactory.getInstance(KeyProperties.KEY_ALGORITHM_EC)
@@ -25,8 +31,8 @@ class VerificationKeys {
         signature.getValidSignaturesForExport(export, signatureListBinary)
             .isEmpty()
             .also {
-                if (it) Timber.d("export is invalid")
-                else Timber.d("export is valid")
+                if (it) Timber.tag(TAG).d("export is invalid")
+                else Timber.tag(TAG).d("export is valid")
             }
     }
 
@@ -43,20 +49,20 @@ class VerificationKeys {
             }
             verified
         }
-        .also { Timber.v("${it.size} valid signatures found") }
+        .also { Timber.tag(TAG).v("${it.size} valid signatures found") }
 
     private fun getKeysForSignatureVerificationFilteredByEnvironment() =
-        BuildConfig.PUB_KEYS_SIGNATURE_VERIFICATION.split(KEY_DELIMITER)
+        environmentSetup.appConfigVerificationKey.split(KEY_DELIMITER)
             .mapNotNull { delimitedString ->
                 Base64.decode(delimitedString, Base64.DEFAULT)
             }.map { binaryPublicKey ->
-            keyFactory.generatePublic(
-                X509EncodedKeySpec(
-                    binaryPublicKey
+                keyFactory.generatePublic(
+                    X509EncodedKeySpec(
+                        binaryPublicKey
+                    )
                 )
-            )
-        }
-            .onEach { Timber.v("$it") }
+            }
+            .onEach { Timber.tag(TAG).v("$it") }
 
     private fun getTEKSignaturesForEnvironment(
         signatureListBinary: ByteArray?
@@ -64,6 +70,6 @@ class VerificationKeys {
         .parseFrom(signatureListBinary)
         .signaturesList
         .asSequence()
-        .onEach { Timber.v(it.toString()) }
+        .onEach { Timber.tag(TAG).v(it.toString()) }
         .mapNotNull { it.signature.toByteArray() }
 }
