@@ -3,8 +3,8 @@ package de.rki.coronawarnapp.worker
 import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import de.rki.coronawarnapp.http.WebRequestBuilder
-import de.rki.coronawarnapp.http.playbook.PlaybookImpl
+import de.rki.coronawarnapp.playbook.Playbook
+import de.rki.coronawarnapp.util.di.AppInjector
 
 /**
  * One time background noise worker
@@ -17,9 +17,8 @@ class BackgroundNoiseOneTimeWorker(
 ) :
     CoroutineWorker(context, workerParams) {
 
-    companion object {
-        private val TAG: String? = BackgroundNoiseOneTimeWorker::class.simpleName
-    }
+    private val playbook: Playbook
+        get() = AppInjector.component.playbook
 
     /**
      * Work execution
@@ -30,8 +29,7 @@ class BackgroundNoiseOneTimeWorker(
         var result = Result.success()
 
         try {
-            PlaybookImpl(WebRequestBuilder.getInstance())
-                .dummy()
+            playbook.dummy()
         } catch (e: Exception) {
             // TODO: Should we even retry here?
             result = if (runAttemptCount > BackgroundConstants.WORKER_RETRY_COUNT_THRESHOLD) {
