@@ -7,16 +7,16 @@ class HTTPVariablesTest {
 
     @Test
     fun getHTTPConnectionTimeout() {
-        Assert.assertEquals(HTTPVariables.getHTTPConnectionTimeout(), 10000L)
+        Assert.assertEquals(HTTPVariables.getHTTPConnectionTimeout(), 20000L)
     }
 
     @Test
     fun getHTTPReadTimeout() {
-        Assert.assertEquals(HTTPVariables.getHTTPReadTimeout(), 10000L)
+        Assert.assertEquals(HTTPVariables.getHTTPReadTimeout(), 20000L)
     }
 
     @Test
     fun getHTTPWriteTimeout() {
-        Assert.assertEquals(HTTPVariables.getHTTPWriteTimeout(), 10000L)
+        Assert.assertEquals(HTTPVariables.getHTTPWriteTimeout(), 20000L)
     }
 }
